requirements.txt
beautifulsoup4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.1
frozendict==2.4.6
idna==3.10
joblib==1.5.1
jupyter
kafka-python
multitasking==0.0.11
nltk==3.9.1
numpy
pandas
peewee==3.18.1
platformdirs==4.3.8
praw
protobuf==6.31.0
pycparser==2.22
pymongo
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
regex==2024.11.6
requests
scikit-learn
scipy==1.15.3
six==1.17.0
soupsieve==2.7
textblob
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance